Define interface for encoder/decoder
