+++ leshan-integration-tests/src/test/java/org/eclipse/leshan/integration/tests/util/QueuedModeLeshanClient.java
@@ -39,8 +39,8 @@ import org.eclipse.leshan.client.californium.impl.SecurityObjectPskStore;
 import org.eclipse.leshan.client.resource.LwM2mObjectEnabler;
 import org.eclipse.leshan.client.servers.BootstrapHandler;
 import org.eclipse.leshan.client.servers.RegistrationEngine;
-import org.eclipse.leshan.core.node.codec.LwM2mNodeDecoder;
-import org.eclipse.leshan.core.node.codec.LwM2mNodeEncoder;
+import org.eclipse.leshan.core.node.codec.DefaultLwM2mNodeDecoder;
+import org.eclipse.leshan.core.node.codec.DefaultLwM2mNodeEncoder;
 import org.eclipse.leshan.util.Validate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -67,7 +67,7 @@ public class QueuedModeLeshanClient implements LwM2mClient {
         private LwM2mObjectEnabler nodeEnabler;
 
         public CustomObjectResource(LwM2mObjectEnabler nodeEnabler) {
-            super(nodeEnabler, bootstrapHandler, new LwM2mNodeEncoder(), new LwM2mNodeDecoder());
+            super(nodeEnabler, bootstrapHandler, new DefaultLwM2mNodeEncoder(), new DefaultLwM2mNodeDecoder());
             this.nodeEnabler = nodeEnabler;
             this.nodeEnabler.setNotifySender(this);
             setObservable(true);
